Protocol change to clarify field names.
Now a field named "frames" will always refer to a Float32Array
whether it is provided as input or arrives as output.

Similarly, a field named "packet" will always refer to a
Uint8Array holding a single Opus encoded packet whether
given as input for decoding or arrives as output of an
encoder.